viewing navigation

+++ src/commands/sudoku/database/saved.ts
@@ -1,6 +1,6 @@
 // view saved games and optionally load them
 
-import { SlashCommandBuilder} from "discord.js";
+import { ButtonInteraction, ChatInputCommandInteraction, ComponentType, SlashCommandBuilder} from "discord.js";
 import SlashCommand from "../../_interface/SlashCommand";
 
 export const saved: SlashCommand = {
@@ -20,26 +20,60 @@ export const saved: SlashCommand = {
         ephemeral: true
       });
     }
-
+    
+    const sudokuSession = interaction.client.sudokuSessions.get(user.id);
+    
+    // verify user is not already viewing games
+    if (sudokuSession.viewing) {
+      return interaction.reply({
+        content: "You are already viewing your games.\nPress exit to stop viewing your games.",
+        ephemeral: true
+      });
+    }
+    
     await interaction.deferReply();
     const message = await interaction.fetchReply();
-
-    const sudokuSession = interaction.client.sudokuSessions.get(user.id);
-
+    
     await sudokuSession.message.delete();
 
     const reply = await sudokuSession.view(message, "Saved");
     
+    const response = await interaction.editReply(reply);
+
+    const collectionFilter = (i: ButtonInteraction) => {return i.user.id === interaction.user.id};
+
+    const collector = response.createMessageComponentCollector({
+      filter: collectionFilter,
+      componentType: ComponentType.Button,
+      time: 3_600_000
+    });
+
+    // listen for button presses in view menu
+    collector.on('collect', async i => {
+      switch (i.customId) {
+        case ("left"):
+          await i.update(await sudokuSession.shiftGame("Saved", "left"));
+          break;
+        case ("right"):
+          await i.update(await sudokuSession.shiftGame("Saved", "right"));
+          break;
+        case ("load"):
+          break;
+        case ("delete"):
+          break;
+        case ("exit"):
+          break;
+      }
+    });
 
     // @[user]'s Saved Games
     // difficulty: [difficulty]
     // [image of game]
-    // game 1 out of ??                       <-- should i do this?
+    // game 1 out of ??                       <-- should i do this? yes
 
     // buttons: [left] [right] [load] [delete] [exit]
 
     // this message will be updated instead of replaced every time a button is pressed.
 
-    await interaction.editReply(reply);
   }
 }
